% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd.R
\name{aggregate_dwd_data}
\alias{aggregate_dwd_data}
\title{Aggregate DWD data}
\usage{
aggregate_dwd_data(dataframe, aggregation_of_time, min_entries = NULL)
}
\arguments{
\item{dataframe}{dataframe returned by [load_dwd_data]{load_dwd_data} or in the same format.}

\item{aggregation_of_time}{character. To what time span should data be aggregated? Options: "hour", "day", "week", "month", "year"}

\item{min_entries}{integer. Minimum number of entries for a single time step in the aggregation to be valid.}
}
\value{
dataframe containing all aggregated stations loaded after one another.
}
\description{
Aggregate downloaded DWD data to what the user likes.
}
\examples{
# create sf_point from central Germany
g <- sf::st_sfc(sf::st_point(x = c(51.163375, 10.447683)), crs = "EPSG:4326")
p <- sf::st_sf(ID = "central_germany", geom = g)

# find 20 nearest dwd stations from central germany that are running til today
dataframe <- find_dwd_stations(p,
                               number_of_stations = 20,
                               mindate = "2024-06-30",
                               res = "hourly",
                               var = "air_temperature",
                               per = "recent",
                               update_rdwd = TRUE)

# download and filter all these stations to desired structure
dwd_raw <- load_dwd_data(dataframe,
                         per = c("recent", "historical"),
                         start = "2022-08-01",
                         end = "2024-07-31")

# aggregate data to monthly means
dwd_aggregated <- aggregate_dwd_data(dwd_raw,
                                     aggregation_of_time = "monthly")

# show data
utils::head(dwd_aggregated)
}
\seealso{
[load_dwd_data]{load_dwd_data}
}
