% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd.R
\name{find_dwd_stations}
\alias{find_dwd_stations}
\title{Search DWD-Stations}
\usage{
find_dwd_stations(
  x,
  number_of_stations,
  mindate,
  res,
  var,
  per,
  update_rdwd = FALSE
)
}
\arguments{
\item{x}{sf_point object with one point.}

\item{number_of_stations}{numeric. Number of stations one wants to collect.}

\item{mindate}{character. Latest date you want to have in the stations. Format: "YYYY-MM-DD"}

\item{res, var, per}{character. Restrictions for dataset type as documented in rdwd::selectDWD(). Each can be a vector of entries.}

\item{update_rdwd}{logical. Use updateRdwd() before execution? Important if one wants the most recent stations from DWD.}
}
\value{
dataframe. A data frame containing all nearby stations defined in number_of_stations
}
\description{
Search n-count of nearby DWD stations from a point in space.
}
\examples{

# create sf_point from central Germany
g <- sf::st_sfc(sf::st_point(x = c(51.163375, 10.447683)), crs = "EPSG:4326")
p <- sf::st_sf(ID = "central_germany", geom = g)

# find 20 nearest dwd stations from central germany that are running til today
dataframe <- find_dwd_stations(p,
                               number_of_stations = 20,
                               mindate = "2024-06-30",
                               res = "hourly",
                               var = "air_temperature",
                               per = "recent",
                               update_rdwd = TRUE)
utils::head(dataframe)
}
