% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd.R
\name{load_dwd_data}
\alias{load_dwd_data}
\title{Load DWD-Stations}
\usage{
load_dwd_data(
  dataframe,
  dir = tempdir(),
  res = NULL,
  var = NULL,
  per = NULL,
  start,
  end
)
}
\arguments{
\item{dataframe}{dataframe returned by \link{find_dwd_data}.}

\item{dir}{character. Directory where DWD_raw_files should be stored. Default = \code{tempdir()}.}

\item{res, var, per}{character. Restrictions for dataset type as documented in rdwd::selectDWD(). Each can be a vector of entries. Per default taken from dataframe.}

\item{start, end}{character. Start and End of Timeseries one wants to load. Format = \code{"yyyy-mm-dd"}}
}
\value{
dataframe containing all stations loaded after one another.
}
\description{
Search n-count of nearby DWD stations from a point in space.
}
\examples{
# create sf_point from central Germany
g <- sf::st_sfc(sf::st_point(x = c(51.163375, 10.447683)), crs = "EPSG:4326")
p <- sf::st_sf(ID = "central_germany", geom = g)

# find 20 nearest dwd stations from central germany that are running til today
dataframe <- find_dwd_stations(p,
                               number_of_stations = 20,
                               mindate = "2024-06-30",
                               res = "hourly",
                               var = "air_temperature",
                               per = "recent",
                               update_rdwd = TRUE)

# download and filter all these stations to desired structure
dwd_raw <- load_dwd_data(dataframe,
                         per = c("recent", "historical"),
                         start = "2022-08-01",
                         end = "2024-07-31")
utils::head(dwd_raw)
}
\seealso{
\link{find_dwd_data}, \code{\link[rdwd:selectDWD]{rdwd::selectDWD()}}
}
