% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Voronoi.R
\name{voronoi}
\alias{voronoi}
\title{Voronoi}
\usage{
voronoi(
  x,
  envelope = NULL,
  values,
  filename,
  main = NULL,
  labels = NULL,
  statistics = FALSE,
  delaunay = FALSE
)
}
\arguments{
\item{x}{To your liking:

- `terra`: SpatVector. A terra SpatVector with spatial points.

- `sf`: Multipolygon. Sf object with spatial points.

- `path`: Character. A path to a vector object with spatial points that can be read by `terra::vect`.}

\item{envelope}{Same as x. Polygon object containing the Area of Interest.}

\item{values}{character. Column Name of your data values.}

\item{filename}{character. Path object. Name of the output file.}

\item{main}{character. Title for your voronoi.}

\item{labels}{Row, Column or Vector of labels in x.}

\item{statistics}{logical. If `TRUE`, basic data statistics will be plotted
next to the diagramm. Default = `FALSE`.}

\item{delaunay}{logical. If `TRUE`, a delauney diagramm will be plotted in
the background. Default = `FALSE`.}
}
\value{
SpatVector. A terra SpatVector with voronoi polygons.
}
\description{
Calculates a voronoi using the terra package
}
\examples{
# load example data
nc <- st_read(system.file("gpkg/nc.gpkg", package="sf"), quiet = TRUE) \%>\%
  summarise()
df <- data.frame(plotID = 1:20,
                 Temperature = runif(n=20, min=5, max=20))
df <- dplyr::mutate(df, st_sample(nc, 20)) \%>\%
  st_as_sf()

# create Voronoi at "WorkingDirectory/Voronoi.pdf"
voronoi(x = df,
        envelope = nc,
        values = "Temperature",
        filename = "Voronoi.pdf",
        main = "Voronoi",
        labels = df$plotID,
        statistics = TRUE,
        delaunay = TRUE)

}
\seealso{
terra, sf
}
